pyproject.toml
[project]
name = "aviation-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "chgwyellow",email = "dinnis1107@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.4.3,<3.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
